Split single Math.h into smaller parts
