+++ NEngine/include/NEngine/Math/Math.h
@@ -1,289 +1,16 @@
 #pragma once
 #include <cstdint>
 #include <cstring>
-#ifdef __cplusplus
 #include <sstream>
-#endif
 
-#if NENGINE_USE_DIRECTXMATH
-#include "DirectXMath.h"
-#endif
+#include "Mat3X3.h"
+#include "Mat4X4.h"
+#include "Vec2D.h"
+#include "Vec3D.h"
+
 
 namespace NEngine {
 namespace Math {
-
-#if NENGINE_USE_DIRECTXMATH
-
-struct Vec2D
-{
-    Vec2D();
-    Vec2D(float x, float y);
-    Vec2D(Vec2D rhs) = default;
-    Vec2D &operator=(Vec2D rhs) = default;
-    float X;
-    float Y;
-};
-
-struct Vec3D
-{
-    Vec3D();
-    Vec3D(float x, float y, float z);
-    Vec3D(Vec3D rhs) = default;
-    Vec3D &operator=(Vec3D rhs) = default;
-    float X;
-    float Y;
-    float Z;
-};
-
-struct Vec4D
-{
-    Vec4D();
-    Vec4D(float x, float y, float z, float w);
-    float X;
-    float Y;
-    float Z;
-    float W;
-};
-
-struct Mat3X3
-{
-};
-
-struct Mat4X4
-{
-};
-
-Mat4X4 MathXMMatrixToMat4X4(DirectX::XMMATRIX xmMat);
-
-#else
-
-struct Vec2D
-{
-    Vec2D()
-        : X{0},
-          Y{0}
-    {
-    }
-
-    Vec2D(float x, float y)
-        : X{x},
-          Y{y}
-    {
-    }
-#ifdef __cplusplus
-    std::string ToString() const;
-
-    Vec2D(const Vec2D &rhs) = default;
-    Vec2D(Vec2D &&rhs) noexcept = default;
-    Vec2D &operator=(const Vec2D &rhs) = default;
-    Vec2D &operator=(Vec2D &&rhs) noexcept = default;
-#endif
-    float X;
-    float Y;
-};
-
-struct Vec3D
-{
-    Vec3D()
-        : X{0},
-          Y{0},
-          Z{0}
-    {
-    }
-
-    Vec3D(float x, float y, float z)
-        : X{x},
-          Y{y},
-          Z{z}
-    {
-    }
-#ifdef __cplusplus
-    std::string ToString() const;
-    Vec3D Cross(const Vec3D &rhs) const;
-    float Dot(const Vec3D &rhs) const;
-    float Length() const;
-    void Normalize();
-    bool operator==(const Vec3D &rhs) const;
-    bool operator!=(const Vec3D &rhs) const;
-
-    Vec3D(const Vec3D &rhs) = default;
-    Vec3D(Vec3D &&rhs) noexcept = default;
-    Vec3D &operator=(const Vec3D &rhs) = default;
-    Vec3D &operator=(Vec3D &&rhs) noexcept = default;
-#endif
-    float X;
-    float Y;
-    float Z;
-};
-
-#ifdef __cplusplus
-Vec3D operator+(const Vec3D &lhs, const Vec3D &rhs);
-Vec3D operator-(const Vec3D &lhs, const Vec3D &rhs);
-Vec3D operator*(const Vec3D &lhs, const float s);
-Vec3D operator*(const float s, const Vec3D &rhs);
-#endif
-
-struct Vec4D
-{
-    Vec4D()
-        : X{0},
-          Y{0},
-          Z{0},
-          W{0}
-    {
-    }
-
-    Vec4D(float x, float y, float z, float w)
-        : X{x},
-          Y{y},
-          Z{z},
-          W{w}
-    {
-    }
-#ifdef __cplusplus
-    Vec4D(const Vec3D &v, float w)
-        : Vec4D(v.X, v.Y, v.Z, w)
-    {
-    }
-
-    Vec4D(const Vec4D &rhs) = default;
-    Vec4D(Vec4D &&rhs) noexcept = default;
-    Vec4D &operator=(const Vec4D &rhs) = default;
-    Vec4D &operator=(Vec4D &&rhs) noexcept = default;
-
-    std::string ToString() const;
-    Vec4D &operator*(const float scalar);
-    Vec4D &operator/(const float scalar);
-    bool operator==(const Vec4D &rhs) const;
-#endif
-    float X;
-    float Y;
-    float Z;
-    float W;
-};
-
-struct Mat3X3
-{
-    Mat3X3()
-    {
-        memset(this, 0, sizeof(Mat3X3));
-    }
-
-    union
-    {
-        struct
-        {
-            float A00, A01, A02;
-            float A10, A11, A12;
-            float A20, A21, A22;
-        };
-
-        Vec3D V[3];
-        float A[3][3];
-    };
-#ifdef __cplusplus
-    std::string ToString() const;
-    float Determinant() const;
-#endif
-};
-
-struct Mat4X4
-{
-    Mat4X4()
-        : Mat4X4(0.0f,
-                 0.0f,
-                 0.0f,
-                 0.0f,
-                 0.0f,
-                 0.0f,
-                 0.0f,
-                 0.0f,
-                 0.0f,
-                 0.0f,
-                 0.0f,
-                 0.0f,
-                 0.0f,
-                 0.0f,
-                 0.0f,
-                 0.0f)
-    {
-    }
-
-    Mat4X4(float a00,
-           float a01,
-           float a02,
-           float a03,
-           float a10,
-           float a11,
-           float a12,
-           float a13,
-           float a20,
-           float a21,
-           float a22,
-           float a23,
-           float a30,
-           float a31,
-           float a32,
-           float a33)
-        : A00{a00},
-          A01{a01},
-          A02{a02},
-          A03{a03},
-          A10{a10},
-          A11{a11},
-          A12{a12},
-          A13{a13},
-          A20{a20},
-          A21{a21},
-          A22{a22},
-          A23{a23},
-          A30{a30},
-          A31{a31},
-          A32{a32},
-          A33{a33}
-    {
-    }
-
-    Mat4X4(const float *data)
-        : Mat4X4(data[0],
-                 data[1],
-                 data[2],
-                 data[3],
-                 data[4],
-                 data[5],
-                 data[6],
-                 data[7],
-                 data[8],
-                 data[9],
-                 data[10],
-                 data[11],
-                 data[12],
-                 data[13],
-                 data[14],
-                 data[15])
-    {
-    }
-
-    union
-    {
-        struct
-        {
-            float A00, A01, A02, A03;
-            float A10, A11, A12, A13;
-            float A20, A21, A22, A23;
-            float A30, A31, A32, A33;
-        };
-
-        Vec4D V[4];
-        float A[4][4];
-    };
-#ifdef __cplusplus
-    std::string ToString() const;
-    bool operator==(const Mat4X4 &rhs) const;
-#endif
-};
-
-#endif
 
 // *** 2D vector math ***
 Vec2D MathVec2DZero(void);
